+++ src/playscreen/panels/NextBlockPanel.java
@@ -10,7 +10,7 @@ import java.awt.Color;
 public class NextBlockPanel extends JPanel {
     private int[][] shape;
     int[] colorTable;
-    private int screenSize;
+    private double screenSize;
     private boolean colorMode;
 
     public void setProps() {
@@ -24,7 +24,7 @@ public class NextBlockPanel extends JPanel {
     public NextBlockPanel(int[][] shape) {
         setProps();
         this.shape = shape;
-        this.setPreferredSize(new java.awt.Dimension(10 * 20 * this.screenSize, 5 * 20 * this.screenSize));
+        this.setPreferredSize(new java.awt.Dimension((int)(10 * 20 * this.screenSize), (int)(5 * 20 * this.screenSize)));
     }
 
     // arr 업데이트 메서드
@@ -36,17 +36,17 @@ public class NextBlockPanel extends JPanel {
     @Override
     protected void paintComponent(Graphics g) {
         super.paintComponent(g);
-        int startX = 4 * 20 * screenSize;
-        int startY = 1 * 20 * screenSize;
+        int startX = (int)(4 * 20 * screenSize);
+        int startY = (int)(1 * 20 * screenSize);
 
         for (int i = 0; i < shape.length; i++) {
             for (int j = 0; j < shape[i].length; j++) {
                 if (shape[i][j] != 0) {
                     int color = colorTable[shape[i][j]];
                     g.setColor(new Color(color)); // 속이 빨간색인 사각형
-                    g.fillRect(startX + j * 20 * screenSize, startY + i * 20 * screenSize, 20 * screenSize, 20 * screenSize);
+                    g.fillRect((int)(startX + j * 20 * screenSize), (int)(startY + i * 20 * screenSize), (int)(20 * screenSize),(int)(20 * screenSize));
                     g.setColor(Color.BLACK); // 테두리는 검은색
-                    g.drawRect(startX + j * 20 * screenSize, startY + i * 20 * screenSize, 20 * screenSize, 20 * screenSize);
+                    g.drawRect((int)(startX + j * 20 * screenSize), (int)(startY + i * 20 * screenSize), (int)(20 * screenSize),(int)(20 * screenSize));
                 }
             }
         }
